Split into two files: cli.py and listener.py
